requirements.txt
mem0ai
streamlit
autogen-agentchat[gemini]~=0.2
google-generativeai
ag2[gemini]
